Move FeatureSearchResultsView and FeatureSearchResults into the Search app
